Flytter spesialiserte funksjoner fra VilkårService til VilkårsvurderingForNyBehandlingService, siden de bare ble brukt derfra
