+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/vilkårsvurdering/VilkårService.kt
@@ -26,9 +26,6 @@ import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.Personopplysning
 import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.PersonopplysningGrunnlagRepository
 import no.nav.familie.ba.sak.kjerne.personident.Aktør
 import no.nav.familie.ba.sak.kjerne.personident.PersonidentService
-import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.VilkårsvurderingMigreringUtils.kopiManglendePerioderFraForrigeVilkårsvurdering
-import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.VilkårsvurderingMigreringUtils.utledPeriodeFom
-import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.VilkårsvurderingMigreringUtils.utledPeriodeTom
 import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.VilkårsvurderingUtils.flyttResultaterTilInitielt
 import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.VilkårsvurderingUtils.lagFjernAdvarsel
 import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.VilkårsvurderingUtils.muterPersonResultatDelete
@@ -309,31 +306,6 @@ class VilkårService(
         }
     }
 
-    fun genererVilkårsvurderingForMigreringsbehandlingMedÅrsakEndreMigreringsdato(
-        behandling: Behandling,
-        forrigeBehandlingSomErVedtatt: Behandling? = null,
-        nyMigreringsdato: LocalDate
-    ): Vilkårsvurdering {
-        val vilkårsvurdering = Vilkårsvurdering(behandling = behandling).apply {
-            personResultater = lagVilkårsvurderingForMigreringsbehandlingMedÅrsakEndreMigreringsdato(
-                this,
-                forrigeBehandlingSomErVedtatt!!,
-                nyMigreringsdato
-            )
-        }
-        return vilkårsvurderingService.lagreNyOgDeaktiverGammel(vilkårsvurdering = vilkårsvurdering)
-    }
-
-    fun genererVilkårsvurderingForHelmanuellMigrering(
-        behandling: Behandling,
-        nyMigreringsdato: LocalDate
-    ): Vilkårsvurdering {
-        val vilkårsvurdering = Vilkårsvurdering(behandling = behandling).apply {
-            personResultater = lagVilkårsvurderingForHelmanuellMigrering(this, nyMigreringsdato)
-        }
-        return vilkårsvurderingService.lagreNyOgDeaktiverGammel(vilkårsvurdering = vilkårsvurdering)
-    }
-
     private fun lagTomVilkårsvurdering(vilkårsvurdering: Vilkårsvurdering): Set<PersonResultat> {
         val personopplysningGrunnlag = hentAktivPersonopplysningGrunnlagThrows(vilkårsvurdering.behandling.id)
 
@@ -454,101 +426,6 @@ class VilkårService(
 
                 val tom: LocalDate? =
                     if (vilkår == Vilkår.UNDER_18_ÅR) person.fødselsdato.plusYears(18).minusDays(1) else null
-
-                val begrunnelse = "Migrering"
-
-                VilkårResultat(
-                    personResultat = personResultat,
-                    erAutomatiskVurdert = false,
-                    resultat = Resultat.OPPFYLT,
-                    vilkårType = vilkår,
-                    periodeFom = fom,
-                    periodeTom = tom,
-                    begrunnelse = begrunnelse,
-                    behandlingId = personResultat.vilkårsvurdering.behandling.id
-                )
-            }.toSortedSet(VilkårResultatComparator)
-
-            personResultat.setSortedVilkårResultater(vilkårResultater)
-
-            personResultat
-        }.toSet()
-    }
-
-    private fun lagVilkårsvurderingForMigreringsbehandlingMedÅrsakEndreMigreringsdato(
-        vilkårsvurdering: Vilkårsvurdering,
-        forrigeBehandlingSomErVedtatt: Behandling,
-        nyMigreringsdato: LocalDate
-    ): Set<PersonResultat> {
-        val forrigeBehandlingsvilkårsvurdering = vilkårsvurderingService
-            .hentAktivForBehandling(forrigeBehandlingSomErVedtatt.id)
-            ?: throw Feil(
-                "Kan ikke kopiere vilkårsvurdering fra forrige behandling ${forrigeBehandlingSomErVedtatt.id}" +
-                    "til behandling ${vilkårsvurdering.behandling.id}"
-            )
-        val personopplysningGrunnlag = hentAktivPersonopplysningGrunnlagThrows(vilkårsvurdering.behandling.id)
-
-        return personopplysningGrunnlag.søkerOgBarn.map { person ->
-            val personResultat = PersonResultat(vilkårsvurdering = vilkårsvurdering, aktør = person.aktør)
-
-            val vilkårTyperForPerson = forrigeBehandlingsvilkårsvurdering.personResultater
-                .single { it.aktør == person.aktør }.vilkårResultater
-                .filter { it.resultat == Resultat.OPPFYLT }
-                .map { it.vilkårType }
-
-            val vilkårResultater = vilkårTyperForPerson.map { vilkår ->
-                val fom = utledPeriodeFom(forrigeBehandlingsvilkårsvurdering, vilkår, person, nyMigreringsdato)
-
-                val tom: LocalDate? =
-                    utledPeriodeTom(forrigeBehandlingsvilkårsvurdering, vilkår, person, fom)
-
-                val begrunnelse = "Migrering"
-
-                VilkårResultat(
-                    personResultat = personResultat,
-                    erAutomatiskVurdert = false,
-                    resultat = Resultat.OPPFYLT,
-                    vilkårType = vilkår,
-                    periodeFom = fom,
-                    periodeTom = tom,
-                    begrunnelse = begrunnelse,
-                    behandlingId = personResultat.vilkårsvurdering.behandling.id
-                )
-            }.toSortedSet(VilkårResultatComparator)
-
-            val manglendePerioder = kopiManglendePerioderFraForrigeVilkårsvurdering(
-                vilkårResultater,
-                forrigeBehandlingsvilkårsvurdering, person
-            )
-            vilkårResultater.addAll(manglendePerioder.map { it.kopierMedParent(personResultat) }.toSet())
-            personResultat.setSortedVilkårResultater(vilkårResultater)
-
-            personResultat
-        }.toSet()
-    }
-
-    private fun lagVilkårsvurderingForHelmanuellMigrering(
-        vilkårsvurdering: Vilkårsvurdering,
-        nyMigreringsdato: LocalDate
-    ): Set<PersonResultat> {
-        val personopplysningGrunnlag = hentAktivPersonopplysningGrunnlagThrows(vilkårsvurdering.behandling.id)
-
-        return personopplysningGrunnlag.søkerOgBarn.map { person ->
-            val personResultat = PersonResultat(vilkårsvurdering = vilkårsvurdering, aktør = person.aktør)
-
-            val vilkårTyperForPerson = Vilkår.hentVilkårFor(person.type)
-            val vilkårResultater = vilkårTyperForPerson.map { vilkår ->
-                val fom = when {
-                    vilkår.gjelderAlltidFraBarnetsFødselsdato() -> person.fødselsdato
-                    nyMigreringsdato.isBefore(person.fødselsdato) -> person.fødselsdato
-                    else -> nyMigreringsdato
-                }
-
-                val tom: LocalDate? = when (vilkår) {
-                    Vilkår.UNDER_18_ÅR -> person.fødselsdato.plusYears(18)
-                        .minusDays(1)
-                    else -> null
-                }
 
                 val begrunnelse = "Migrering"
